Combine Pypp test cases
